pyproject.toml
[project]
name = "multi-system-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.13",
    "langchain-openai>=1.1.12",
    "langfuse>=4.0.1",
    "langsmith>=0.7.22",
    "python-dotenv>=1.2.2",
]
